fix(ci): fix Go json-rpc id type assertion in server tests
After readResponse was changed to use json.Unmarshal (so error codes come
back as float64), the id field also comes back as float64 from JSON.
Two tests compared response["id"] against the untyped constant 1, which Go
defaults to int — interface{float64(1)} != interface{int(1)} evaluates to
true, failing the assertion.

Fix: compare against float64(1) explicitly in both affected tests.

+++ code/packages/go/json-rpc/json_rpc_test.go
@@ -561,7 +561,7 @@ func TestServer_DispatchesRequestToHandler(t *testing.T) {
 	server.Serve()
 
 	response := readResponse(out)
-	if response["id"] != 1 {
+	if response["id"] != float64(1) {
 		t.Errorf("id: got %v", response["id"])
 	}
 	if response["result"] != float64(3) {
@@ -686,7 +686,7 @@ func TestServer_NullResultIsValid(t *testing.T) {
 	server.Serve()
 
 	response := readResponse(out)
-	if response["id"] != 1 {
+	if response["id"] != float64(1) {
 		t.Errorf("id: got %v", response["id"])
 	}
 	// result key should be present (with null value)
